flash scripts: simplify release directory common code
Instead of duplicating some defaults in flash_m2k.sh & flash_pluto.sh, do
this in the common_flash.sh code.

Hardcode the release directories in the flash_m2k.sh & flash_pluto.sh
scripts. The aim is to provide simplicity to the users, so we'll harcode
the release directories, and provide scripts to update the hardcoded
release dirs.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>