setup_env.sh: wrap commands that need sudo with the sudo cmd
If we want to setup non-sudo stuff, then running the whole script as sudo
would mess with permissions.
This should localize the need of sudo to a few commands that need it.

Also, since this script should be run once [at setup], we're removing the
file checks.
They don't seem to be reliable anyway.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>