config/xilinx-tcl.cfg: drop m2k_reset_jtag  func
It's pretty difficult to control JTAG init/reset via OpenOCD, because the
`init` function will scan the JTAG chain, and will probably error out.

So, we won't use a special reset function for JTAG.
We'll define a nSRST signal, which OpenOCD will use to reset/init the board
on startup, which is pretty elegant.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>