config/ftdi4232,xilinx-tcl: implement flashing with special JTAG boot mode
For digilent-hs2.cfg  the pin definitions are dummy so that the script
works without complaining about undefined vars.

While for ftdi4232.cfg they've been updated to the current board setup that
will be used for programming Pluto/M2k.

There will be 2 nPOWER signals, so nPOWER is now nPOWER1.

Added nTRST, which is nPOWER1 + nPOWER2.
This signal is used by OpenOCD to reset/init the board before doing the
JTAG scan chain.
The JTAG_BMODE signal [JTAG boot] is set to high when the board is
initialized; that way, the board will boot into JTAG mode, and we can
disable that pin during the load_uboot() func in `xilinx-tcl.cfg`.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>