build: add utility for bitbanging some pins
This will allow some flexibility for toggling the power pins and the JTAG
boot mode pin.

Using OpenOCD here is tricky as OpenOCD will try a JTAG scan by default.
That scan can probably be disabled, but it requires extra config files, and
that may be a bit cumbersome to maintain.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>