common_flash.sh: implement power cycling for ft4232h board
This uses the ft4232_pin_ctrl utility to toggle the power pins.
Admittedly, this will toggle all pins on channel A to low first, and then
toggle just the power pins back up.
But, at this point in time, JTAG should be done, so there should be no
overlap/problem.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>