ft4232h_pin_ctrl: handle bug for reading channel B in subsequent reads
Seems that after reading the first sample, in the next read there is no
data for channel B, and only from the 3rd sampling onwards we have data for
channel B as well.
No idea yet if this is a HW bug.
After checking configuration data, it all looks fine.

The STATUSEN and CRCEN bits seem to be fine, and the first sample reading
is fine.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>