production_pluto: implement the preflash test (voltage measurements)
This has been largely adapted from the previous production-test scripts,
but uses more shell constructs to achieve the same result.

A `production_test.sh` script wraps the `common_preflash.sh` script.
The `common_preflash.sh` script tries to be generic as much as possible to
be able to extend this to M2k at a later point in time.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>