ft4232h_pin_ctrl: initialize pin values to 0
The compiler does not seem to give an error here, but masking values on
undefined stack memory is bad.
And it's problematic on ARM, which seems to init init stack variables with
undefined values [unless specifically specified].

This is not an issue on Intel.
Since doing this, the buffer has been extended to 3 bytes.
The 3rd byte is the direction of the pins.
Might as well set this, to avoid any other [potentially] undefined
behavior.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>