lib: organize common files under the lib folder
They're piling up, so put common stuff into a lib folder.
This is somewhat of a dumb split, since now the `common_flash.sh` script
won't be able to run on it's own.

But running `flash_pluto.sh` works fine.
This requires a bit of shell magic to adapt, but it could be that we may
want this behavior, to prevent errors/goofs.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>