lib/utils.sh,ft4232h_pin_ctrl: implement basic SPI read for EEPROM
This means that the old `spi` mode now becomes spi-adc, to make room for
mode `spi-eeprom`.

The EEPROM has a chip-select on channel D [of FTDI], so we need to
select/control that externally.

This change implements the basic commands for:
* reading/writing data to the EEPROM
* reading/writing data to the Status Register

In the shell script, only the R/W data commands are implemented.
But the caller has to be aware of which page the EEPROM writes to.

For the current EEPROM [of 512 bytes] a page size of 16 bytes.
If the user wants to write more than 16 bytes, he/she will have to update
the address if it jumps across a page.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>