reset_adc.sh: drop script and reset logic
Was added during AD7616 development.
But the ADC will always be reset, before doing any measurements to make
sure it's a valid state before measuring anything.

It's also desired to this, in case USB has any glitches which could put the
ADC in a weird state that would require reset anyway.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>