measure.sh: support passing device or voltage ranges for ADC
Some channels require a higher precision, which means lowering the voltage
range on certain channels.
For M2k, these values are in the `config/m2k/values.sh` file.

So, the measure.sh script has been extended to take a 2nd param which can
be either the device name [m2k for the moment] or a list of voltage ranges
for each channel.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>