eeprom_cfg.sh: rework eeprom save/load and vars
We need to save some variables in the EEPROM that are per-channel.
So, we need to extend the old mechanism and add VOFF0A, VFF0B, VGAIN0A,
VGAIN0B env vars which get saved into EEPROM.

Then, we construct the set for all 16 channels.
VGAIN & VOFF are applied as defaults, and VGAINXY & VOFFXY are applied on
top.

This also means that the entries in the EEPROM need to be dynamic [up to
32 entries max]. Some VOFF & VGAIN [per channel] entries need not be
specified.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>