config/m2k/scopy.js: implement support for pressing buttons to continue in step 9
We cannot wait for the `wait_pins.sh` script inside the scopy.js script.
That will hang the main program loop in Scopy, because and it will block
the entire process.

So, a basic/simple IPC is required with the wait_pins.sh going into
background, and outputing to a file which scopy.js will poll until
output is printed.

We also need to force-fully terminate any wait_pins.sh processes that
are lingering, in case something goes wrong and processes stall or
become zombie.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>