ft4232h_pin_ctrl: implement no-samples and no-samples-avg suboptions
For debugging and maybe also for the real case, we may want to see which
measurement value comes up more often.

So, to do this, we need an option to print all samples that we measure
with the tool, to be able to do this sort of processing at a reasonable
speed.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>