m2k power calibration: add special script/rules for user power calibration
The calibration of the power supplies is a bit special, and requires as
higher precision. And given that the ADC in the system doesn't always
read correct values, we need to some democratic elections of
measurements and try to filter out the values that are out of range.

Typically, when values go out of range, they go wayyyyy out of range.

This can be due to electrical issues, or driver problems, or USB or
FTDI.
The best we can do is try to filter out the best stuff through this
multitude of potential error sources.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>