config/m2k/postflash.sh: split scopy.js script in 2 parts
We only need the GUI for the trimmers part.
The first steps until that part can be run with the `--nogui` option.

In the 2nd part Scopy is started up with the GUI visible.
There seems to be an issue with loading the GUI hidden and then making
it visible.
Elements don't seem to paint properly, hence the reason for splitting
this script.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>