autosave_logs.sh,setup_env.sh: add auto-save logs mechanism
It's a simple dumb loop with 5 second intervals that polls to see if there
is an `autosave_logs` folder in /media folder (where USB & SD cards get
mounted).

Once there is one, it picks up the first one and saves all log files there.

The `setup_env.sh` script will install this support to run this when user
logs in.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>