setup_env.sh: disallow root as caller of this script
The intent is to run this script with a normal root.
setup_env.sh will make sure [and ask] for sudo access later.
The idea is that some things are needed to be installed in the user's home
folder, so running this script as root complicates things, as home folder
may be /root

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>