lib/flash.sh,flash.exp: remove ttyUSB requirement & related code
This code is a left-over from when we used the UART interface to check
whether device is in boot-mode, or maybe some other thing.

The TTY-USB interface is not required for any flashing procedure, so remove
the code entirely.
Also, rename cmd.exp -> flash.exp to make it clearer what that file does.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>