lib/flash.sh: pass a boot-mode flag to the exp script
Apparently the JTAG_BMODE variable does not get passed properly from the
config script (of OpenOCD) to the xilinx-tcl.cfg script.

That means that the GPIO for boot-mode won't be disabled on boot, which
won't allow the device to go into  DFU mode.

To fix this, it's passed as an argument to the lib/flash.exp script.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>