setup_env.sh: setup zerotier vpn
This sets up the ZeroTier VPN service, which was recommended by the guys at
Seeed. This will help us connect to the jigs in China. The security aspect
of the service is questionable, but the ease of use is very good, so we
need to take care of the first part.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>