pluto: rework extra environment generation
Because, we needed to enable JFFS2 on the Pluto, we also needed to disable
4K sector access, which meant writing the extra environment variables to
the flash doesn't work with the old production code.

So, now we're bypassing that a bit, and writing 64K sectors (standard
access size), which works even with 4K sector-access disabled.

This means using mtd_debug to erase & write it, and also we need to
generate the entire information to write to flash to the same location we
used to write before (with 4K sector access enable).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>